Prevent piston exploits (dupes, bedrock breaking)
Scheduled piston events will use the direction of the piston at the time
of the event creation rather than execution. This prevents the piston
from being replaced with a different facing one mid-tick, as is used
often in bedrock breaking.

Known forms of duping from pistons are fixed, mainly TNT and carpet
duping. This is not fixing the root of the problem that is causing such
dupes to happen, but rather is an explicit check against TNT so that a
block 36 entity isn't created for TNT that doesn't exist in the world
anymore.